working on the UI
