+++ src/pages/App.tsx
@@ -9,7 +9,7 @@ import FarmingV2 from "./FarmingV2";
 import Fonts from "./../theme/fonts";
 import AppWrapper from "./../components/AppWrapper";
 import Navbar from "./../components/Navbar";
-
+import ConfirmModal from "../components/modals/ConfirmModal";
 export default function App() {
   return (
     <Suspense fallback={null}>
@@ -29,6 +29,9 @@ export default function App() {
             />
             <Route exact path="/farming" component={FarmingV1} />
             <Route exact path="/farming-V2" component={FarmingV2} />
+            <ConfirmModal 
+            title="confirm"
+            />
             <Route path="/">
               <Redirect to="/swap" />
             </Route>
